requirements.txt
bleak>=0.21.0
aiohttp>=3.9.0
Pillow>=10.0.0
